requirements.txt
python-dotenv==1.0.0
crewai>=0.114.0
unstructured>=0.16.4
langchain-community>=0.3.22
langchain-mcp-adapters>=0.0.1a2
langgraph>=0.2.12
langchain-openai>=0.1.2 